move command structs out of prerunner
